# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/workspaces/current": {
      "get": {
        "operationId": "WorkspacesController_current",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "trial",
                        "active",
                        "suspended",
                        "closed"
                      ]
                    },
                    "country": {
                      "nullable": true,
                      "type": "string"
                    },
                    "defaultLocale": {
                      "type": "string"
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "dataRegion": {
                      "type": "string",
                      "enum": [
                        "eu",
                        "in",
                        "us"
                      ]
                    },
                    "trialEndsAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "createdVia": {
                      "type": "string",
                      "enum": [
                        "platform",
                        "partner",
                        "self_service"
                      ]
                    },
                    "logoFileId": {
                      "nullable": true,
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "branding": {
                      "type": "object",
                      "properties": {
                        "accentColor": {
                          "description": "Accent colour for reports and emails (#RRGGBB)",
                          "nullable": true,
                          "type": "string",
                          "pattern": "^#[0-9A-Fa-f]{6}$"
                        },
                        "reportFooter": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 500
                        }
                      },
                      "additionalProperties": false
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "slug",
                    "status",
                    "country",
                    "defaultLocale",
                    "timezone",
                    "dataRegion",
                    "trialEndsAt",
                    "createdVia",
                    "logoFileId",
                    "branding",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      },
      "patch": {
        "operationId": "WorkspacesController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "country": {
                    "nullable": true,
                    "type": "string",
                    "pattern": "^[A-Z]{2}$"
                  },
                  "defaultLocale": {
                    "type": "string",
                    "pattern": "^[a-z]{2,3}(-[A-Z]{2})?$"
                  },
                  "timezone": {
                    "type": "string",
                    "maxLength": 64
                  },
                  "branding": {
                    "type": "object",
                    "properties": {
                      "accentColor": {
                        "description": "Accent colour for reports and emails (#RRGGBB)",
                        "nullable": true,
                        "type": "string",
                        "pattern": "^#[0-9A-Fa-f]{6}$"
                      },
                      "reportFooter": {
                        "nullable": true,
                        "type": "string",
                        "maxLength": 500
                      }
                    },
                    "additionalProperties": false
                  },
                  "logoFileId": {
                    "nullable": true,
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "trial",
                        "active",
                        "suspended",
                        "closed"
                      ]
                    },
                    "country": {
                      "nullable": true,
                      "type": "string"
                    },
                    "defaultLocale": {
                      "type": "string"
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "dataRegion": {
                      "type": "string",
                      "enum": [
                        "eu",
                        "in",
                        "us"
                      ]
                    },
                    "trialEndsAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "createdVia": {
                      "type": "string",
                      "enum": [
                        "platform",
                        "partner",
                        "self_service"
                      ]
                    },
                    "logoFileId": {
                      "nullable": true,
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "branding": {
                      "type": "object",
                      "properties": {
                        "accentColor": {
                          "description": "Accent colour for reports and emails (#RRGGBB)",
                          "nullable": true,
                          "type": "string",
                          "pattern": "^#[0-9A-Fa-f]{6}$"
                        },
                        "reportFooter": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 500
                        }
                      },
                      "additionalProperties": false
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "slug",
                    "status",
                    "country",
                    "defaultLocale",
                    "timezone",
                    "dataRegion",
                    "trialEndsAt",
                    "createdVia",
                    "logoFileId",
                    "branding",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
    "/v1/workspaces/current/entitlements": {
      "get": {
        "operationId": "WorkspacesController_entitlements",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Plan, modules, limits and current usage",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plan": {
                      "type": "string"
                    },
                    "modules": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "gap",
                          "materiality",
                          "actions",
                          "surveys",
                          "supply_chain",
                          "ranking",
                          "frameworks",
                          "ai",
                          "carbon"
                        ]
                      }
                    },
                    "limits": {
                      "type": "object",
                      "properties": {
                        "companies": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "users": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "projectsPerYear": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "clientWorkspaces": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "aiMonthlyUsd": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        }
                      },
                      "additionalProperties": false
                    },
                    "trialEndsAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "usage": {
                      "type": "object",
                      "properties": {
                        "companies": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "users": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991,
                          "description": "Active members + pending invitations"
                        },
                        "clientWorkspaces": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "companies",
                        "users",
                        "clientWorkspaces"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "plan",
                    "modules",
                    "limits",
                    "trialEndsAt",
                    "usage"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
    "/v1/workspaces/current/partner-grants": {
      "get": {
        "operationId": "WorkspacesController_partnerGrants",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Partners with an active grant to this workspace",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "partnerWorkspaceId": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "partnerName": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "partnerWorkspaceId",
                          "partnerName",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
    "/v1/workspaces/current/partner-grants/{id}": {
      "delete": {
        "operationId": "WorkspacesController_revokePartnerGrant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
...
  },
  "info": {
    "title": "ResiliSense API",
    "description": "ResiliSense 2.0 REST API. Specs: docs/revamp/modules. Errors are RFC 9457 problem+json.",
    "version": "2.0.0-alpha",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}